% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binFunc.R
\name{mybin}
\alias{mybin}
\title{Binomial Experiment Function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{How many experiments to run}

\item{n}{Number of trials}

\item{p}{Probability of success}
}
\description{
Runs a binomial experiment
}
\examples{
mybin(iter=1000,n=18, p=0.3)
}
