% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(x, iter = 10000, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{x}{The input data}

\item{iter}{number of samples to make}

\item{fun}{Which function to use in apply}

\item{alpha}{The confidence interval}

\item{cx}{Font size}

\item{...}{Extra settings}
}
\value{
List of the confidence interval, function used and the orignal data set
}
\description{
Bootstrap Function
}
\examples{
set.seed(39)
sam=rnorm(25,mean=25,sd=10)
myboot2(sam)
}
