% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson Distribution sampler}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{Number of samples taken per experiment}

\item{iter}{Number of experiments to run}

\item{lambda}{lambda for poisson}

\item{...}{extra selection options}
}
\value{
Null
}
\description{
Poisson Distribution sampler
}
\examples{
mycltp(n=2, iter=10000,lambda=4)
mycltp(n=3, iter=10000,lambda=4)
mycltp(n=5, iter=10000,lambda=4)
mycltp(n=10,iter=10000,lambda=4)
}
