% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myhyper.R
\name{myhyper}
\alias{myhyper}
\title{myhyper funcation}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{fds}

\item{N}{fds}

\item{r}{fds}

\item{n}{fds}
}
\description{
fds
}
\examples{
myhyper()
}
