% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{The max likehood function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Function to use for making the matrix}

\item{x}{The values for the columns}

\item{param}{The values for the rows}

\item{...}{Extra stuff for plot}
}
\value{
info about the results
}
\description{
The max likehood function
}
\examples{
logbin=function(x, param){
log(dbinom(x, 20, param))
}
mymaxlik(logbin, c(3,3,4,3,4,5,5,4), seq(0,1,length=1000))
}
