% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve function}
\usage{
myncurve(mu = 0, sigma = 1, a = 0.5)
}
\arguments{
\item{mu}{Mean of the function}

\item{sigma}{Standard deviation of the function}

\item{a}{Right endpoint}
}
\description{
Plots a normal distribution curve that shows how much
and where a lower tail prop is
}
\examples{
myncurve()
}
