% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{My Sample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{Number of pulls per set}

\item{iter}{Number of time to run the experiment}

\item{time}{does nothing}
}
\description{
My Sample
}
\examples{
mysample(5)
}
