% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{nticket function for project 1}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the plane}

\item{gamma}{Probability of plane being overbook}

\item{p}{Probability of a show}
}
\value{
A list with the details the of experiment
}
\description{
The nticket function for project 1. Given an number of seats, the probability
the plane is overbook and the probability of a show, it will output the
number of seats that can be sold without causing pain and graphs.
}
\examples{
ntickets(N=400,gamma = 0.02, p = 0.95)
}
