% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Pareto graphing funcation}
\usage{
pareto(x, mn = "Pareto barplot", ...)
}
\arguments{
\item{x}{the data in vector form}

\item{mn}{graph title}

\item{...}{extra stuff}
}
\description{
Pareto graphing funcation
}
\examples{
freq=c(15,8,63,20)
RL=c("None","Both","LegsO","WheelsO")
l=rep(RL,freq)
pareto(l)
}
