% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMTSS_old.R
\name{RMTSS_old}
\alias{RMTSS_old}
\title{RSS MSS TSS Function}
\usage{
RMTSS_old(model, val)
}
\arguments{
\item{model}{A fit model}

\item{val}{The data that is being checked}
}
\description{
Prints the RSS, MSS, TSS and MSS / TSS
}
\examples{
library(MATH4753NATH24)
}
